+++ src/onboardingSteps/onboarding-steps.service.ts
@@ -4,15 +4,29 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { OnboardingStep } from './entities/onboarding-step.entity';
 import { CreateOnboardingStepDto } from './dtos/create-onboarding-step.dto';
 import { UpdateOnboardingStepDto } from './dtos/update-onboarding-step.dto';
+import { ApiKey } from 'src/auth/entities/api-key.entity';
+import {
+  mapOnboardingStepEntityToDto,
+  mapOnboardingStepsToDtos,
+} from './mappers/mapEntitiesToDto';
 
 @Injectable()
 export class OnboardingStepsService {
   constructor(
     @InjectRepository(OnboardingStep)
     private readonly onboardingStepRepository: Repository<OnboardingStep>,
+    @InjectRepository(ApiKey)
+    private readonly apiKeyRepository: Repository<ApiKey>,
   ) {}
 
-  async create(createOnboardingStepDto: CreateOnboardingStepDto) {
+  async create(
+    headerApiKey: string,
+    createOnboardingStepDto: CreateOnboardingStepDto,
+  ) {
+    const savedApiKey = await this.apiKeyRepository.findOne({
+      where: { key: headerApiKey },
+    });
+
     const createdCountryCode = this.onboardingStepRepository.create({
       title: createOnboardingStepDto.title,
       subtitle: createOnboardingStepDto.subtitle,
@@ -20,19 +34,27 @@ export class OnboardingStepsService {
       imageUrl: createOnboardingStepDto.imageUrl,
       details: createOnboardingStepDto.details,
       order: createOnboardingStepDto.order,
+      createdBy: savedApiKey,
     });
 
-    return await this.onboardingStepRepository.save(createdCountryCode);
+    const savedOnboardingStep =
+      await this.onboardingStepRepository.save(createdCountryCode);
+
+    return mapOnboardingStepEntityToDto(savedOnboardingStep);
   }
 
   async findAll() {
-    return await this.onboardingStepRepository.find();
+    const allOnboardingSteps = await this.onboardingStepRepository.find();
+
+    return mapOnboardingStepsToDtos(allOnboardingSteps);
   }
 
   async findAllById(onboardingStepIds: string[]) {
-    return await this.onboardingStepRepository.findBy({
+    const allOnboardingStepsById = await this.onboardingStepRepository.findBy({
       id: In(onboardingStepIds),
     });
+
+    return mapOnboardingStepsToDtos(allOnboardingStepsById);
   }
 
   async findOne(id: string) {
@@ -44,11 +66,18 @@ export class OnboardingStepsService {
       throw new NotFoundException('Onboarding step was not found');
     }
 
-    return onboardingStep;
+    return mapOnboardingStepEntityToDto(onboardingStep);
   }
 
-  async update(id: string, updateCountryCodeDto: UpdateOnboardingStepDto) {
+  async update(
+    headerApiKey: string,
+    id: string,
+    updateCountryCodeDto: UpdateOnboardingStepDto,
+  ) {
     const onboardingStep = await this.findOne(id);
+    const savedApiKey = await this.apiKeyRepository.findOne({
+      where: { key: headerApiKey },
+    });
 
     Object.assign(onboardingStep, {
       title: updateCountryCodeDto.title || onboardingStep.title,
@@ -58,14 +87,27 @@ export class OnboardingStepsService {
       imageUrl: updateCountryCodeDto.imageUrl || onboardingStep.imageUrl,
       details: updateCountryCodeDto.details || onboardingStep.details,
       order: updateCountryCodeDto.order || onboardingStep.order,
+      updatedBy: savedApiKey,
     });
 
-    return await this.onboardingStepRepository.save(onboardingStep);
+    const updatedStep =
+      await this.onboardingStepRepository.save(onboardingStep);
+
+    return mapOnboardingStepEntityToDto(updatedStep);
   }
 
   async remove(id: string) {
-    const onboardingStep = await this.findOne(id);
+    const onboardingStep = await this.onboardingStepRepository.findOneBy({
+      id,
+    });
 
-    return await this.onboardingStepRepository.remove(onboardingStep);
+    if (!onboardingStep) {
+      throw new NotFoundException('Onboarding step was not found');
+    }
+
+    const removedOnboardingStep =
+      await this.onboardingStepRepository.remove(onboardingStep);
+
+    return mapOnboardingStepEntityToDto(removedOnboardingStep);
   }
 }
