+++ src/utils/fake-forms-service.util.ts
@@ -6,11 +6,17 @@ import { FormsService } from 'src/forms/forms.service';
 import { getFakeEntityRepository } from './fake-repository.util';
 import { isFilledArray } from './isFilledArray';
 import { mapFormFields } from 'src/forms/mappers/form-fields.mapper';
+import { ApiKey } from 'src/auth/entities/api-key.entity';
+import {
+  mapFormEntityToDto,
+  mapFormsToDtos,
+} from 'src/forms/mappers/mapEntitiesToDto';
 
 const fakeFormRepository = getFakeEntityRepository<Form>();
+const fakeApiKeyRepository = getFakeEntityRepository<ApiKey>();
 
 export const fakeFormsService: Partial<FormsService> = {
-  create: async (createFormDto: CreateFormDto) => {
+  create: async (headerApiKey: string, createFormDto: CreateFormDto) => {
     const mappedFields = mapFormFields(createFormDto.fields);
 
     if (!isFilledArray(mappedFields)) {
@@ -19,30 +25,33 @@ export const fakeFormsService: Partial<FormsService> = {
       );
     }
 
+    const savedApiKey = await fakeApiKeyRepository.findOne({
+      where: { key: headerApiKey },
+    });
+
     const createdForm = fakeFormRepository.create({
       name: createFormDto.name,
       fields: mappedFields,
+      createdBy: savedApiKey,
     });
 
-    const savedForm = fakeFormRepository.save(createdForm);
+    const savedForm = await fakeFormRepository.save(createdForm);
 
-    return await savedForm;
+    return mapFormEntityToDto(savedForm);
   },
   findAll: async () => {
-    return await fakeFormRepository.find();
+    const allForms = await fakeFormRepository.find();
+
+    return mapFormsToDtos(allForms);
   },
   findOne: async (id: string) => {
-    if (!id) {
-      return null;
-    }
-
     const foundForm = await fakeFormRepository.findOneBy({ id });
 
     if (!foundForm) {
       throw new NotFoundException('Form was not found');
     }
 
-    return foundForm;
+    return mapFormEntityToDto(foundForm);
   },
 
   // update: async (id: string, updatePolicyDto: UpdateFormDto) => {
@@ -57,8 +66,23 @@ export const fakeFormsService: Partial<FormsService> = {
   //   return await fakeFormRepository.save(foundForm);
   // },
   remove: async (id: string) => {
-    const foundForm = await fakeFormsService.findOne(id);
+    const foundForm = await fakeFormRepository.findOneBy({ id });
 
-    return await fakeFormRepository.remove(foundForm);
+    if (!foundForm) {
+      throw new NotFoundException('Form was not found');
+    }
+    const removedForm = await fakeFormRepository.remove(foundForm);
+
+    return mapFormEntityToDto(removedForm);
   },
 };
+
+Object.assign(fakeFormsService, {
+  initApiKey: async () => {
+    const createdApiKey = await fakeApiKeyRepository.create({
+      key: process.env.API_KEY,
+      appName: 'Dev Testing',
+    });
+    await fakeApiKeyRepository.save(createdApiKey);
+  },
+});
